+++ draft-js-dnd-plugin/src/utils/file.js
@@ -1,7 +1,7 @@
 // Check if drag event contains files (not text)
 export function containsFiles(event) {
   if (event.dataTransfer.types) {
-    for (let i = 0; i < event.dataTransfer.types.length; i++) {
+    for (let i = 0; i < event.dataTransfer.types.length; i += 1) {
       if (event.dataTransfer.types[i] === 'Files') {
         return true;
       }
@@ -14,11 +14,11 @@ export function containsFiles(event) {
 // Read file contents intelligently as plain text/json, image as dataUrl or
 // anything else as binary
 export function readFile(file) {
-  return new Promise(resolve => {
+  return new Promise((resolve) => {
     const reader = new FileReader();
 
     // This is called when finished reading
-    reader.onload = event => {
+    reader.onload = (event) => {
       // Return an array with one image
       resolve({
         // These are attributes like size, name, type, ...
@@ -30,9 +30,6 @@ export function readFile(file) {
 
         // This is the files content as base64
         src: event.target.result,
-
-        // No URL, since nothing on server
-        url: null,
       });
     };