+++ draft-js-emoji-plugin/src/index.js
@@ -1,19 +1,22 @@
+import { Map, List } from 'immutable';
+
+import keys from 'lodash.keys';
+import decorateComponentWithProps from 'decorate-component-with-props';
+import { EditorState } from 'draft-js';
 import Emoji from './Emoji';
 import EmojiSuggestions from './EmojiSuggestions';
 import EmojiSuggestionsPortal from './EmojiSuggestionsPortal';
 import emojiStrategy from './emojiStrategy';
 import emojiSuggestionsStrategy from './emojiSuggestionsStrategy';
-import decorateComponentWithProps from 'decorate-component-with-props';
-import { Map } from 'immutable';
 import emojiStyles from './emojiStyles.css';
 import emojiSuggestionsStyles from './emojiSuggestionsStyles.css';
 import emojiSuggestionsEntryStyles from './emojiSuggestionsEntryStyles.css';
 import attachImmutableEntitiesToEmojis from './modifiers/attachImmutableEntitiesToEmojis';
 import defaultPositionSuggestions from './utils/positionSuggestions';
-import { EditorState } from 'draft-js';
+import emojiList from './utils/emojiList';
 
 const defaultImagePath = '//cdn.jsdelivr.net/emojione/assets/svg/';
-const cacheBustParam = '?v=2.1.2';
+const cacheBustParam = '?v=2.2.6';
 
 // TODO activate/deactivate different the conversion or search part
 
@@ -50,7 +53,7 @@ const createEmojiPlugin = (config = {}) => {
   };
 
   let searches = Map();
-  let escapedSearch = undefined;
+  let escapedSearch;
   let clientRectFunctions = Map();
 
   const store = {
@@ -91,7 +94,11 @@ const createEmojiPlugin = (config = {}) => {
     theme = defaultTheme,
     positionSuggestions = defaultPositionSuggestions,
     imagePath = defaultImagePath,
+    priorityList,
   } = config;
+
+  // if priorityList is configured in config then set priorityList
+  if (priorityList) emojiList.setPriorityList(priorityList);
   const emojiSearchProps = {
     ariaProps,
     cacheBustParam,
@@ -100,6 +107,7 @@ const createEmojiPlugin = (config = {}) => {
     theme,
     store,
     positionSuggestions,
+    shortNames: List(keys(emojiList.list)),
   };
   return {
     EmojiSuggestions: decorateComponentWithProps(EmojiSuggestions, emojiSearchProps),