+++ draft-js-table-plugin/src/components/nested-editor.js
@@ -1,8 +1,7 @@
 import React, { Component } from 'react';
 import { EditorState, ContentState, convertFromRaw, convertToRaw } from 'draft-js';
-import ReactDOM from 'react-dom';
 
-export default PluginEditor => class NestedEditor extends Component {
+export default (PluginEditor) => class NestedEditor extends Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -13,24 +12,23 @@ export default PluginEditor => class NestedEditor extends Component {
   }
 
   componentDidMount() {
-    this.DOMNode = ReactDOM.findDOMNode(this.refs.editor);
-    this.DOMNode.addEventListener('mousedown', this.mouseDown, false);
-    this.DOMNode.addEventListener('keydown', this.stopPropagation, false);
+    this.editor.addEventListener('mousedown', this.mouseDown, false);
+    this.editor.addEventListener('keydown', this.stopPropagation, false);
   }
 
   componentWillUnmount() {
-    this.DOMNode.removeEventListener('mousedown', this.listener, false);
-    this.DOMNode.removeEventListener('keydown', this.stopPropagation, false);
+    this.editor.removeEventListener('mousedown', this.listener, false);
+    this.editor.removeEventListener('keydown', this.stopPropagation, false);
   }
 
-  onChange = editorState => {
+  onChange = (editorState) => {
     const { readOnly } = this.props;
     if (readOnly) return;
     this.setState({ editorState });
     this.props.onChange(convertToRaw(editorState.getCurrentContent()));
   }
 
-  mouseDown = event => {
+  mouseDown = (event) => {
     const { readOnly, setFocus } = this.props;
     event.stopPropagation();
     if (readOnly === false) {
@@ -40,7 +38,7 @@ export default PluginEditor => class NestedEditor extends Component {
     setFocus();
   }
 
-  stopPropagation = event => {
+  stopPropagation = (event) => {
     if (event.keyCode === 38) {
       event.stopPropagation();
     } else if (event.keyCode === 40) {
@@ -55,7 +53,13 @@ export default PluginEditor => class NestedEditor extends Component {
     const { readOnly } = this.props;
 
     return (
-      <PluginEditor {...this.props} ref="editor" editorState={editorState} onChange={this.onChange} readOnly={readOnly} />
+      <PluginEditor
+        {...this.props}
+        ref={(element) => { this.editor = element; }}
+        editorState={editorState}
+        onChange={this.onChange}
+        readOnly={readOnly}
+      />
     );
   }
 };