+++ draft-js-toolbar-plugin/src/decorators/hover-toolbar.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 
 // Get a component's display name
-const getDisplayName = WrappedComponent => {
+const getDisplayName = (WrappedComponent) => {
   const component = WrappedComponent.WrappedComponent || WrappedComponent;
   return component.displayName || component.name || 'Component';
 };
@@ -10,7 +10,7 @@ const getDisplayName = WrappedComponent => {
 let number = 0;
 
 // HoverToolbar decorator will render a toolbar on hovering the WrappedComponent
-export default (defaultTheme, toolbarStore) => WrappedComponent => class HoverToolbarDecorator extends Component {
+export default (defaultTheme, toolbarStore) => (WrappedComponent) => class HoverToolbarDecorator extends Component {
   // Statics
   static displayName = `HoverToolbar(${getDisplayName(WrappedComponent)})`;
   static pluginOptions = WrappedComponent.pluginOptions;
@@ -25,9 +25,11 @@ export default (defaultTheme, toolbarStore) => WrappedComponent => class HoverTo
   // Bind listeners on mount
   componentDidMount() {
     // Set this.number to a unique value
-    this.number = number = number++;
+    number += 1;
+    this.number = number;
 
     // Get this domNode
+    // eslint-disable-next-line react/no-find-dom-node
     const element = ReactDOM.findDOMNode(this);
     if (!element) {
       return;
@@ -51,7 +53,7 @@ export default (defaultTheme, toolbarStore) => WrappedComponent => class HoverTo
   }
 
   // Show the toolbar
-  showToolbar = event => {
+  showToolbar = (event) => {
     this.doRenderToolbar(true);
     if (event) event.stopPropagation();
   };
@@ -82,7 +84,7 @@ export default (defaultTheme, toolbarStore) => WrappedComponent => class HoverTo
   };
 
   // Render the actual toolbar
-  doRenderToolbar = active => {
+  doRenderToolbar = (active) => {
     const props = {
       ...this.props,
       actions: [...(this.props.actions || []), ...(this.componentActions || [])],
@@ -99,7 +101,7 @@ export default (defaultTheme, toolbarStore) => WrappedComponent => class HoverTo
     }
   }
 
-  addActions = actions => {
+  addActions = (actions) => {
     this.componentActions = actions;
   }